Feat: update verify(), evaluate() fn

+++ foaks/linear_gkr/src/circuit_fast_track/mod.rs
@@ -7,7 +7,7 @@ pub struct Gate {
     pub ty: u32,
     pub u: usize,
     pub v: usize,
-    pub src: Vec<i32>,
+    pub src: Vec<usize>,
     pub weight: Vec<FieldElement>,
     pub parameter_length: usize,
 }
